+++ CHANGELOG.md
@@ -6,9 +6,13 @@ All notable changes to [camunda-bpmn-js-behaviors](https://github.com/camunda/ca
 
 ___Note:__ Yet to be released changes appear here._
 
+## 1.12.1
+
+* `FIX`: fix crash when call activity without extension elements is used in Camunda 7 ([#114](https://github.com/camunda/camunda-bpmn-js-behaviors/pull/114))
+
 ## 1.12.0
 
-* `FEAT`: add `OrderExtensionElementsBehavior` for Camunda Platform ([#113](https://github.com/camunda/camunda-bpmn-js-behaviors/pull/113))
+* `FEAT`: add `OrderExtensionElementsBehavior` for Camunda 7 ([#113](https://github.com/camunda/camunda-bpmn-js-behaviors/pull/113))
 * `FEAT`: add `CallActivityVariablesPropagationBehavior` ([#112](https://github.com/camunda/camunda-bpmn-js-behaviors/pull/112))
 
 ## 1.11.3